add unit test for innertx
